+++ src/events/HumanRemoveEvent.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace beeaz\humannpc\event;
+namespace BeeAZ\HumanNPC\events;
 
 use pocketmine\entity\Entity;
 use pocketmine\event\entity\EntityEvent;
 use pocketmine\player\Player;
 
-class HumanNPCRemoveEvent extends EntityEvent {
+class HumanRemoveEvent extends EntityEvent {
     private ?Player $remover;
 
     public function __construct(Entity $entity, ?Player $remover = null) {
